+++ include/utils.h
@@ -20,12 +20,12 @@ char *read_line(void);
 void check_ninja_binary_format(FILE *fp, char *arg);
 void check_ninja_version(FILE *fp, char *arg);
 void read_instructions_into_ir(FILE *fp);
-size_t check_ninja_instruction_count(FILE *fp);
-size_t check_ninja_variable_count(FILE *fp);
-Bytecode seek_file(FILE *fp, int offset);
-ObjRef new_composite_object(unsigned int num_obj_refs);
-size_t get_obj_ref_bytes(ObjRef obj_ref);
-size_t get_obj_ref_size(ObjRef obj_ref);
+unsigned check_ninja_instruction_count(FILE *fp);
+unsigned check_ninja_variable_count(FILE *fp);
+Bytecode seek_file(FILE *fp, unsigned offset);
+ObjRef new_composite_object(unsigned num_obj_refs);
+unsigned get_obj_ref_bytes(ObjRef obj_ref);
+unsigned get_obj_ref_size(ObjRef obj_ref);
 void set_broken_heart(ObjRef obj_ref);
 void set_forward_pointer(ObjRef obj_ref, unsigned forward_pointer);
